pyproject.toml
[project]
name = "niemand_server"
version = "0.1.0"
description = "Niemand server"
dependencies = [
    "fastapi~=0.109.0",
    "pydantic~=2.5.3",
    "uvicorn[standard]~=0.25.0",
    "aiohttp~=3.9.1",
    "gender-determinator~=0.2.1",
    "beautifulsoup4~=4.12.2",
    "openai~=1.51.0",
    "azure-cognitiveservices-speech~=1.34.0",
    "pvporcupine~=2.1.4",
    "caldav~=1.3.9",
    "icalendar~=6.0.0",
    "dependency-injector~=4.42.0",
    "setuptools~=75.1.0",
    "python-dateutil~=2.9.0",
    "apscheduler~=3.10.4"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["src/**"]

[tool.hatch.build.targets.wheel]
include = ["src/**"]

[tool.hatch.envs.default]
python = "3.12"
path = ".venv"